+++ Tests/KeyboardKitTests/Actions/StandardKeyboardActionBehaviorTests.swift
@@ -29,7 +29,7 @@ class StandardKeyboardActionBehaviorTests: QuickSpec {
         
         describe("preferred keyboard type after gesture on action") {
             
-            func result(for gesture: KeyboardGesture, on action: KeyboardAction, context: KeyboardContext) -> KeyboardType {
+            func result(for context: KeyboardContext, after gesture: KeyboardGesture, on action: KeyboardAction) -> KeyboardType {
                 behavior.preferredKeyboardType(for: context, after: gesture, on: action)
             }
             
@@ -37,76 +37,76 @@ class StandardKeyboardActionBehaviorTests: QuickSpec {
                 proxy.documentContextBeforeInput = "Hello!"
                 proxy.autocapitalizationType = .allCharacters
                 context.keyboardType = .alphabetic(.lowercased)
-                expect(result(for: .tap, on: .character("i"), context: context)).to(equal(.alphabetic(.uppercased)))
+                expect(result(for: context, after: .tap, on: .character("i"))).to(equal(.alphabetic(.uppercased)))
             }
             
             it("is context preferred keyboard type if shift is double-tapped too slowly") {
                 context.keyboardType = .alphabetic(.uppercased)
                 behavior.lastShiftCheck = .distantPast
-                expect(result(for: .tap, on: .shift(currentState: .uppercased), context: context)).to(equal(.alphabetic(.uppercased)))
+                expect(result(for: context, after: .tap, on: .shift(currentState: .uppercased))).to(equal(.alphabetic(.uppercased)))
             }
             
             it("is context preferred keyboard type if non-shift is double-tapped") {
                 context.keyboardType = .alphabetic(.uppercased)
-                expect(result(for: .tap, on: .command, context: context)).to(equal(.alphabetic(.lowercased)))
+                expect(result(for: context, after: .tap, on: .command)).to(equal(.alphabetic(.lowercased)))
             }
             
             it("is context preferred keyboard type if current keyboard type is not upper-cased") {
                 context.keyboardType = .alphabetic(.lowercased)
-                expect(result(for: .tap, on: .command, context: context)).to(equal(.alphabetic(.lowercased)))
+                expect(result(for: context, after: .tap, on: .command)).to(equal(.alphabetic(.lowercased)))
             }
             
             it("is caps-locked if shift is double-tapped quickly") {
                 context.keyboardType = .alphabetic(.uppercased)
-                expect(result(for: .tap, on: .shift(currentState: .uppercased), context: context)).to(equal(.alphabetic(.capsLocked)))
+                expect(result(for: context, after: .tap, on: .shift(currentState: .uppercased))).to(equal(.alphabetic(.capsLocked)))
             }
         }
         
         
         describe("should end sentence after gesture on action") {
             
-            func result(for gesture: KeyboardGesture, on action: KeyboardAction, context: KeyboardContext) -> Bool {
+            func result(for context: KeyboardContext, after gesture: KeyboardGesture, on action: KeyboardAction) -> Bool {
                 behavior.shouldEndSentence(for: context, after: gesture, on: action)
             }
             
             it("is only true for space after a previous space") {
                 proxy.documentContextBeforeInput = "hej  "
-                expect(result(for: .tap, on: .space, context: context)).to(beTrue())
-                expect(result(for: .tap, on: .character(" "), context: context)).to(beTrue())
-                expect(result(for: .tap, on: .command, context: context)).to(beFalse())
+                expect(result(for: context, after: .tap, on: .space)).to(beTrue())
+                expect(result(for: context, after: .tap, on: .character(" "))).to(beTrue())
+                expect(result(for: context, after: .tap, on: .command)).to(beFalse())
                 proxy.documentContextBeforeInput = "hej. "
-                expect(result(for: .tap, on: .space, context: context)).to(beFalse())
-                expect(result(for: .tap, on: .character(" "), context: context)).to(beFalse())
+                expect(result(for: context, after: .tap, on: .space)).to(beFalse())
+                expect(result(for: context, after: .tap, on: .character(" "))).to(beFalse())
                 proxy.documentContextBeforeInput = "hej.  "
-                expect(result(for: .tap, on: .space, context: context)).to(beFalse())
-                expect(result(for: .tap, on: .character(" "), context: context)).to(beFalse())
+                expect(result(for: context, after: .tap, on: .space)).to(beFalse())
+                expect(result(for: context, after: .tap, on: .character(" "))).to(beFalse())
             }
         }
         
         describe("should switch to preferred keyboard type after gesture on action") {
             
-            func result(for gesture: KeyboardGesture, on action: KeyboardAction, context: KeyboardContext) -> Bool {
+            func result(for context: KeyboardContext, after gesture: KeyboardGesture, on action: KeyboardAction) -> Bool {
                 behavior.shouldSwitchToPreferredKeyboardType(for: context, after: gesture, on: action)
             }
             
             it("is true if the action is shift") {
-                expect(result(for: .tap, on: .shift(currentState: .capsLocked), context: context)).to(beTrue())
-                expect(result(for: .tap, on: .shift(currentState: .lowercased), context: context)).to(beTrue())
-                expect(result(for: .tap, on: .shift(currentState: .uppercased), context: context)).to(beTrue())
+                expect(result(for: context, after: .tap, on: .shift(currentState: .capsLocked))).to(beTrue())
+                expect(result(for: context, after: .tap, on: .shift(currentState: .lowercased))).to(beTrue())
+                expect(result(for: context, after: .tap, on: .shift(currentState: .uppercased))).to(beTrue())
             }
             
             it("is true is the current keyboard type differs from the preferred one") {
                 proxy.documentContextBeforeInput = "Hello!"
                 proxy.autocapitalizationType = .allCharacters
                 context.keyboardType = .alphabetic(.lowercased)
-                expect(result(for: .tap, on: .character("i"), context: context)).to(beTrue())
+                expect(result(for: context, after: .tap, on: .character("i"))).to(beTrue())
             }
             
             it("is false is the current keyboard type is the same as the preferred one") {
                 proxy.documentContextBeforeInput = "Hello!"
                 proxy.autocapitalizationType = .none
                 context.keyboardType = .alphabetic(.lowercased)
-                expect(result(for: .tap, on: .character("i"), context: context)).to(beFalse())
+                expect(result(for: context, after: .tap, on: .character("i"))).to(beFalse())
             }
         }
     }